# prisma/swagger.yaml
openapi: "3.0.0"
paths:
...
  /api/blog/images:
    post:
      operationId: "BlogController_uploadImage"
      summary: "Upload a blog image"
      parameters: []
      responses:
        201:
          description: ""
          content:
            application/json:
              schema:
                properties:
                  url:
                    type: "string"
      tags:
        - "blog"
      security:
        - bearer: []
  /api/blog/categories/vehicle:
    get:
      operationId: "BlogController_getVehicleCategories"
      summary: "Get all vehicle entry categories"
      parameters: []
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/VehicleCategoryDto"
      tags:
        - "blog"
      security:
        - bearer: []
  /api/blog:
    get:
      operationId: "BlogController_getEntries"
      summary: "Get all blog entries for authenticated user"
      parameters:
        - name: "category"
          required: false
          in: "query"
          schema:
            enum:
              - "PERSONAL"
              - "VEHICLE"
            type: "string"
        - name: "car_id"
          required: false
          in: "query"
          schema:
            type: "number"
        - name: "vehicle_category"
          required: false
          in: "query"
          schema:
            enum:
              - "REPAIR"
              - "SERVICE_VISIT"
              - "TRIP"
              - "FUEL"
              - "UPGRADE"
              - "INSPECTION"
              - "BREAKDOWN"
              - "OTHER"
            type: "string"
        - name: "search"
          required: false
          in: "query"
          description: "Search in title and content"
          schema:
            type: "string"
        - name: "is_pinned"
          required: false
          in: "query"
          schema:
            type: "boolean"
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/BlogEntryDto"
      tags:
        - "blog"
      security:
        - bearer: []
    post:
      operationId: "BlogController_createEntry"
      summary: "Create a blog entry"
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBlogEntryDto"
      responses:
        201:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogEntryDto"
      tags:
        - "blog"
      security:
        - bearer: []
  /api/blog/{id}:
    get:
      operationId: "BlogController_getEntry"
      summary: "Get a single blog entry"
      parameters:
        - name: "id"
          required: true
          in: "path"
          schema:
            type: "string"
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogEntryDto"
        404:
          description: "Not found"
      tags:
        - "blog"
      security:
        - bearer: []
    put:
      operationId: "BlogController_updateEntry"
      summary: "Update a blog entry"
      parameters:
        - name: "id"
          required: true
          in: "path"
          schema:
            type: "string"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBlogEntryDto"
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogEntryDto"
      tags:
        - "blog"
      security:
        - bearer: []
    delete:
      operationId: "BlogController_deleteEntry"
      summary: "Delete a blog entry"
      parameters:
        - name: "id"
          required: true
          in: "path"
          schema:
            type: "string"
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogEntryDto"
      tags:
        - "blog"
      security:
        - bearer: []
  /api/blog/{id}/pin:
    patch:
      operationId: "BlogController_togglePin"
      summary: "Toggle pin on a blog entry"
      parameters:
        - name: "id"
          required: true
          in: "path"
          schema:
            type: "string"
      responses:
        200:
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogEntryDto"
      tags:
        - "blog"
      security:
        - bearer: []
...
  /api/auth/google/redirect:
    get:
      operationId: "AuthController_googleAuthRedirect"
      summary: "Handle Google OAuth callback"
      parameters: []
      responses:
        302:
          description: "Authenticated and redirected to frontend."
        401:
          description: "Google authentication failed."
      tags:
        - "auth"
  /api/auth/refresh:
    post:
      operationId: "AuthController_refreshAccessToken"
      summary: "Refresh access token"
      parameters: []
      responses:
        200:
          description: "New access token issued."
        401:
          description: "Refresh token missing or invalid."
      tags:
        - "auth"
  /api/auth/me:
    get:
      operationId: "AuthController_getProfile"
      summary: "Get authenticated user profile"
      parameters: []
      responses:
        200:
          description: "User profile returned."
        401:
          description: "Invalid or missing access token."
      tags:
        - "auth"
  /api/auth/logout:
    post:
      operationId: "AuthController_logout"
      summary: "Logout user and invalidate tokens"
      parameters: []
      responses:
        200:
          description: "Logged out successfully."
        401:
          description: "Not authenticated."
      tags:
        - "auth"
info:
  title: "History Auto Utility API"
  description: "API documentation for History Auto Utility"
  version: "1.0"
  contact: {}
...
servers:
  - url: "http://localhost:3000"
    description: "development"
components:
  securitySchemes:
    bearer:
      scheme: "bearer"
      bearerFormat: "JWT"
      type: "http"
  schemas:
...
    VehicleCategoryDto:
      type: "object"
      properties:
        value:
          type: "string"
          enum:
            - "REPAIR"
            - "SERVICE_VISIT"
            - "TRIP"
            - "FUEL"
            - "UPGRADE"
            - "INSPECTION"
            - "BREAKDOWN"
            - "OTHER"
        label:
          type: "string"
          example: "Repair"
      required:
        - "value"
        - "label"
    BlogTagDto:
      type: "object"
      properties:
        id:
          type: "number"
          example: 1
        label:
          type: "string"
          example: "oil change"
        color:
          type: "string"
          example: "blue"
      required:
        - "id"
        - "label"
        - "color"
    BlogImageDto:
      type: "object"
      properties:
        id:
          type: "number"
          example: 1
        url:
          type: "string"
          example: "/uploads/blog/photo.jpg"
      required:
        - "id"
        - "url"
    BlogEntryDto:
      type: "object"
      properties:
        id:
          type: "number"
          example: 1
        user_id:
          type: "number"
          example: 1
        category:
          type: "string"
          enum:
            - "PERSONAL"
            - "VEHICLE"
          example: "VEHICLE"
        title:
          type: "string"
          example: "Oil service"
        slug:
          type: "string"
          example: "oil-service-2026"
          nullable: true
        excerpt:
          type: "string"
          example: "Short summary."
          nullable: true
        date:
          type: "string"
          example: "2026-01-15T00:00:00.000Z"
        content:
          type: "string"
          example: "Changed oil and filter."
        content_json:
          type: "object"
          example:
            type: "doc"
            content: []
          nullable: true
        status:
          type: "string"
          enum:
            - "DRAFT"
            - "PUBLISHED"
          example: "DRAFT"
        published_at:
          type: "string"
          example: "2026-01-15T10:00:00.000Z"
          nullable: true
        is_pinned:
          type: "boolean"
          example: false
        cover_gradient:
          type: "string"
          example: "linear-gradient(135deg, #667eea 0%, #764ba2 100%)"
          nullable: true
        cover_image_url:
          type: "string"
          example: "/uploads/blog/cover.jpg"
          nullable: true
        car_id:
          type: "number"
          example: 1
          nullable: true
        car_name:
          type: "string"
          example: "BMW 520i"
          nullable: true
        vehicle_category:
          type: "string"
          enum:
            - "REPAIR"
            - "SERVICE_VISIT"
            - "TRIP"
            - "FUEL"
            - "UPGRADE"
            - "INSPECTION"
            - "BREAKDOWN"
            - "OTHER"
          nullable: true
        km:
          type: "number"
          example: 184300
          nullable: true
        price:
          type: "number"
          example: 450
          nullable: true
        tags:
          type: "array"
          items:
            $ref: "#/components/schemas/BlogTagDto"
        images:
          type: "array"
          items:
            $ref: "#/components/schemas/BlogImageDto"
        created_at:
          type: "string"
          example: "2026-01-15T10:00:00.000Z"
        updated_at:
          type: "string"
          example: "2026-01-15T10:00:00.000Z"
      required:
        - "id"
        - "user_id"
        - "category"
        - "title"
        - "slug"
        - "excerpt"
        - "date"
        - "content"
        - "content_json"
        - "status"
        - "published_at"
        - "is_pinned"
        - "cover_gradient"
        - "cover_image_url"
        - "car_id"
        - "car_name"
        - "vehicle_category"
        - "km"
        - "price"
        - "tags"
        - "images"
        - "created_at"
        - "updated_at"
    CreateBlogTagDto:
      type: "object"
      properties:
        label:
          type: "string"
          example: "oil change"
        color:
          type: "string"
          example: "blue"
      required:
        - "label"
        - "color"
    CreateBlogEntryDto:
      type: "object"
      properties:
        category:
          type: "string"
          enum:
            - "PERSONAL"
            - "VEHICLE"
        title:
          type: "string"
          example: "Oil service"
        slug:
          type: "string"
          example: "oil-service-2026"
        excerpt:
          type: "string"
          example: "Short summary of the entry."
        date:
          type: "string"
          example: "2026-01-15"
        content:
          type: "string"
          example: "Changed oil and filter."
        content_json:
          type: "object"
          example:
            type: "doc"
            content: []
          description: "Tiptap JSON document"
        status:
          type: "string"
          enum:
            - "DRAFT"
            - "PUBLISHED"
          default: "DRAFT"
        is_pinned:
          type: "boolean"
          example: false
        cover_gradient:
          type: "string"
          example: "linear-gradient(135deg, #667eea 0%, #764ba2 100%)"
        cover_image_url:
          type: "string"
          example: "/uploads/blog/cover.jpg"
        car_id:
          type: "number"
          example: 1
        vehicle_category:
          type: "string"
          enum:
            - "REPAIR"
            - "SERVICE_VISIT"
            - "TRIP"
            - "FUEL"
            - "UPGRADE"
            - "INSPECTION"
            - "BREAKDOWN"
            - "OTHER"
        km:
          type: "number"
          example: 184300
        price:
          type: "number"
          example: 450
        tags:
          type: "array"
          items:
            $ref: "#/components/schemas/CreateBlogTagDto"
        images:
          description: "Array of gallery image URLs"
          type: "array"
          items:
            type: "string"
      required:
        - "category"
        - "title"
        - "date"
    UpdateBlogEntryDto:
      type: "object"
      properties:
        category:
          type: "string"
          enum:
            - "PERSONAL"
            - "VEHICLE"
        title:
          type: "string"
          example: "Oil service"
        slug:
          type: "string"
          example: "oil-service-2026"
        excerpt:
          type: "string"
        date:
          type: "string"
          example: "2026-01-15"
        content:
          type: "string"
        content_json:
          type: "object"
          description: "Tiptap JSON document"
        status:
          type: "string"
          enum:
            - "DRAFT"
            - "PUBLISHED"
        is_pinned:
          type: "boolean"
        cover_gradient:
          type: "string"
        cover_image_url:
          type: "string"
        car_id:
          type: "number"
        vehicle_category:
          type: "string"
          enum:
            - "REPAIR"
            - "SERVICE_VISIT"
            - "TRIP"
            - "FUEL"
            - "UPGRADE"
            - "INSPECTION"
            - "BREAKDOWN"
            - "OTHER"
        km:
          type: "number"
        price:
          type: "number"
        tags:
          type: "array"
          items:
            $ref: "#/components/schemas/CreateBlogTagDto"
        images:
          type: "array"
          items:
            type: "string"